Use the Partitioner on the C++ to build graph.
On the Python side, we have assumed 2 boxes so far. This commit
adds two functions in the shared library so that the Python code
can make use of the C++ partitioner to build the execution graph.

Also of note in this commit is the use of `max_calls` in the
remote function. There seems to be runtime issue (memory/handle
leaks?) that indeterministically hangs or fails the simulation.
This option allows us to restart the worker fresh.