Replace all floats from thesis-bench-floats to doubles
The kernel was not exactly the same as the thesis-bench-doubles.
There was use of internal CUDA helper_math calls such as `length`,
`dot`, and the operators for float3.
Also no SimParams object in the thesis-bench-doubles branch,
nor GetMooreBoxIndexes (i.e. still had triple for loop over all 27
boxes).

We should introduces this branch as an another improvement over
the thesis-bench-doubles branch, due to these small changes
(depending on benchmark results).